+++ django-data/image/biosample/tasks/submission.py
@@ -8,7 +8,7 @@ Created on Thu Jul 18 14:14:06 2019
 
 import redis
 import traceback
-import pyUSIrest.client
+import pyUSIrest.usi
 
 from celery import chord
 from celery.utils.log import get_task_logger
@@ -109,7 +109,7 @@ class SubmissionHelper():
 
     def read_token(self):
         """Read token from REDIS database and set root attribute with a
-        pyUSIrest.client.Root instance
+        pyUSIrest.usi.Root instance
 
         Returns:
             str: the read token"""
@@ -135,7 +135,7 @@ class SubmissionHelper():
         self.auth = get_auth(token=self.token)
 
         logger.debug("getting biosample root")
-        self.root = pyUSIrest.client.Root(auth=self.auth)
+        self.root = pyUSIrest.usi.Root(auth=self.auth)
 
         return self.token
 
@@ -183,7 +183,7 @@ class SubmissionHelper():
         """Create a new USI submission object
 
         Returns:
-            :py:class:`pyUSIrest.client.Submission` a pyUSIrest submission
+            :py:class:`pyUSIrest.usi.Submission` a pyUSIrest submission
             object
         """