+++ HISTORY.rst
@@ -15,7 +15,7 @@ TODO
 Features
 ^^^^^^^^
 
-* upgrade ``pyUSIrest`` dependency
+* pinning ``pyUSIrest`` dependency to ``v0.3.0``
 * export *BioSamples IDs* as a CSV from *Edit Submission view*
 * force excel date cells into ``datetime.date`` objects
 * remove duplicated ``animal_age_at_collection`` column from excel template